PET Assembler Development Systems
This is an continuation of my earlier work to reverse engineer
the Commodore Macro Development System's (MADS) Assembler.

MADS is based on a tool set that started on the PET. Included
are versions for the PET 2 and the PET 4.

The goal is to produce source code that can be complied using the
tools in the native environment using VICE.

This is a work in progress based on source code generated by my
Disassembler running on the C64.

Denton Marlowe
April 2018